+++ app-nexus-network/app-nexus-network-htb-exports.js
@@ -15,8 +15,3 @@
 window.pbjs = window.pbjs || {};
 window.pbjs.renderApnxNetAd = SpaceCamp.services.RenderService.renderDfpAd.bind(null, 'AppNexusNetworkHtb');
 //? }
-
-if (__directInterface.Layers.PartnersLayer.Partners.AppNexusNetworkHtb) {
-    shellInterface.AppNexusNetworkHtb = shellInterface.AppNexusNetworkHtb || {};
-    shellInterface.AppNexusNetworkHtb.adResponseCallback = __directInterface.Layers.PartnersLayer.Partners.AppNexusNetworkHtb.adResponseCallback;
-}