+++ share-through/share-through-htb-system-tests.js
@@ -10,7 +10,7 @@ function getStatsId() {
 function getBidRequestRegex() {
   return {
     method: "GET",
-    urlRegex: /.*btlr\.sharethrough\.com\/t6oivhQt\/.*/
+    urlRegex: /https:\/\/btlr\.sharethrough\.com\/t6oivhQt\/.*/
   };
 }
 
@@ -39,26 +39,29 @@ function getConfig() {
 
 function validateBidRequest(request) {
   var queryObj = request.query;
-
   expect(queryObj.bidId).toBeDefined();
   expect(queryObj.placement_key).toEqual("abc123");
   expect(queryObj.instant_play_capable).toBeDefined();
   expect(queryObj.hbSource).toEqual("indexExchange");
-  expect(queryObj.hbVersion).toEqual("2.1.2");
+  expect(queryObj.hbVersion).toEqual("2.3.0");
   expect(queryObj.cbust).toBeDefined();
   expect(queryObj.consent_required).toEqual("false");
+  expect(queryObj.secure).toBeDefined();
 }
 
 function validateBidRequestWithPrivacy(request) {
   var queryObj = request.query;
-
   expect(queryObj.consent_required).toEqual("true");
   expect(queryObj.consent_string).toEqual("TEST_GDPR_CONSENT_STRING");
 }
 
+function validateBidRequestWithUspapi(request) {
+  var queryObj = request.query;
+  expect(queryObj.us_privacy).toEqual("TEST_USPAPI_CONSENT_STRING");
+}
+
 function validateBidRequestWithAdSrvrOrg(request) {
   var queryObj = request.query;
-
   expect(queryObj.ttduid).toEqual("TEST_ADSRVR_ORG_STRING");
 }
 
@@ -68,7 +71,7 @@ function getValidResponse(request, creative) {
     creatives: [
       {
         adm: creative,
-        cpm: 2,
+        cpm: 2
       }
     ]
   });
@@ -99,6 +102,7 @@ module.exports = {
   getConfig: getConfig,
   validateBidRequest: validateBidRequest,
   validateBidRequestWithPrivacy: validateBidRequestWithPrivacy,
+  validateBidRequestWithUspapi: validateBidRequestWithUspapi,
   validateBidRequestWithAdSrvrOrg: validateBidRequestWithAdSrvrOrg,
   getValidResponse: getValidResponse,
   validateTargeting: validateTargeting,