+++ triple-lift/triple-lift-htb-exports.js
@@ -1,0 +1,9 @@
+//? if(FEATURES.GPT_LINE_ITEMS) {
+shellInterface.TripleLiftHtb = {
+    render: SpaceCamp.services.RenderService.renderDfpAd.bind(null, 'TripleLiftHtb')
+};
+
+shellInterface.TPL = {
+    render: SpaceCamp.services.RenderService.renderDfpAd.bind(null, 'TripleLiftHtb')
+};
+//? }