+++ pub-matic/CHANGES.md
@@ -2,3 +2,6 @@
 # 2.1.3
 - Initial adapter implementation in New repo
 - SSL Changes- Making secure call to PubMatic server 
+
+# 2.1.4
+- Add CCPA support